data-processing/requirements.txt
# Core NLP Libraries
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# HTTP Requests
requests>=2.31.0
httpx>=0.24.0

# Environment Management
python-dotenv>=1.0.0

# JSON Processing
jsonschema>=4.17.0

# Text Processing
regex>=2023.6.3
unidecode>=1.3.6

# Machine Learning (Optional)
transformers>=4.30.0
torch>=2.0.0

# Vector Operations (Optional)
faiss-cpu>=1.7.4

# Async Processing
asyncio>=3.4.3
aiohttp>=3.8.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# LangChain & OpenAI
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0

# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0